Superclass for integration tests
